+++ changelog.md
@@ -14,6 +14,9 @@
  * Fix the broken `Functor`, `Applicative`, and `Monad` instances for
    `Array` and `SmallArray`.
 
+
+ * Add `PrimUnlifted` instances for `TVar` and `MVar`.
+
  * Use `compareByteArrays#` for the `Eq` and `Ord` instances of
    `ByteArray` when building with GHC 8.4 and newer.